Crate calc_common

Contains shared types for the calc-rs project.

Note: the Character and MetaButton types are laid out to match the following 5x5 matrix hardwired into the current keyboard prototype. The Button ID for a given Button is row_num * 5 + col_num.

                Buttons
     0    1      2      3       4             
   .----.------.------.-------.-------.   External buttons
 0 | 0  | 1    | 2    | 3     | 4     |       25    26
   :----+------+------+-------+-------:    .-----.------.
 1 | 5  | 6    | 7    | 8     | 9     |    | Alt | Menu |
   :----+------+------+-------+-------:    '-----'------'
 2 | +  | -    | x    | /     | ^     |
   :----+------+------+-------+-------:
 3 | (  | )    | .    | Left  | Right |
   :----+------+------+-------+-------:
 4 | Up | Down | Back | Enter | Clear |
   '----'------'------'-------'-------'

Modules

prelude

util

Structs

Button

A compact, one byte representation of a button. Can be either a Character or MetaButton

Point

Enums

Character

Represents a character that can be put into an expression.

ExtendedCharacter

Represents a charactor or symbol that cannot be put into an expression

GetButton

Returned by Button::get. Represents a Character or a MetaButton

GetDrawableCharacter

Represents either a Character or an ExtendedCharacter. This is returned by DrawableCharacter::get

MetaButton

A button that isn’t a Character (eg Enter)

Constants

NUM_BUTTON_CHARS

The number of characters with their own dedicated buttons

NUM_CHARS

NUM_ECHARS

NUM_META_BUTTONS

Unions

DrawableCharacter

Represents either a Character or an ExtendedCharacter